feat: 위치 컴포넌트 모둘화 리팩토링
