+++ components/viewTrip/tripPlan/day/Day.tsx
@@ -1,43 +1,22 @@
-import { planState } from "@/types/viewTrip/viewTrip";
+import { DayProps, planState } from "@/types/viewTrip/viewTrip";
 import styles from "./styles";
 import { ScrollView, Text, TouchableOpacity, View } from "react-native";
 import { AntDesign, MaterialIcons } from "@expo/vector-icons";
 import Locations from "../location/Location";
 import { useState } from "react";
 import AddLocationModal from "../location/modal/AddLocationModal";
-
-interface DayProps {
-  day: number;
-  destination: string;
-  locations: planState["locations"];
-  setDays: React.Dispatch<React.SetStateAction<planState[]>>;
-}
+import DayHeader from "./DayHeader/DayHeader";
 
 const Day: React.FC<DayProps> = ({ day, destination, locations, setDays }) => {
   const [isAddModalVisible, setIsAddModalVisible] = useState(false);
   return (
     <View style={styles.dayContainer} key={day}>
-      <View style={styles.dayHeader}>
-        <View style={styles.dayInfo}>
-          <Text style={styles.dayText}>{day} 일차</Text>
-          <View style={styles.destinationContainer}>
-            <MaterialIcons name="place" size={20} color="#4E5968" />
-            <Text style={styles.destinationText}>{destination}</Text>
-            <View style={styles.badge}>
-              <Text style={styles.badgeText}>{locations.length}곳</Text>
-            </View>
-          </View>
-        </View>
-        <View style={{ display: "flex", alignItems: "center" }}>
-          <TouchableOpacity
-            style={styles.addButton}
-            onPress={() => setIsAddModalVisible(true)}
-          >
-            <Text style={styles.addButtonText}>새로운 장소 추가</Text>
-          </TouchableOpacity>
-        </View>
-      </View>
-
+      <DayHeader
+        day={day}
+        destination={destination}
+        locations={locations}
+        setIsAddModalVisible={setIsAddModalVisible}
+      />
       <ScrollView
         showsVerticalScrollIndicator={false}
         contentContainerStyle={styles.scrollContent}
