+++ components/viewTrip/tripPlan/location/LocationCardFooter/styles.tsx
@@ -1,0 +1,22 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  footerContainer: {
+    padding: 10,
+  },
+  footerGradient: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "center",
+    padding: 10,
+    borderRadius: 16,
+    gap: 8,
+  },
+  footerText: {
+    fontSize: 16,
+    fontWeight: "600",
+    color: "#fff",
+    letterSpacing: -0.3,
+  },
+});
+export default styles;
