+++ types/viewTrip/viewTrip.ts
@@ -107,3 +107,30 @@ export interface useTripPlanProps {
   setDays: React.Dispatch<React.SetStateAction<planState[]>>;
   setCurrentIndex: React.Dispatch<React.SetStateAction<number>>;
 }
+export interface DayHeaderProps {
+  day: number;
+  destination: string;
+  locations: planState["locations"];
+  setIsAddModalVisible: React.Dispatch<React.SetStateAction<boolean>>;
+}
+export interface DayProps {
+  day: number;
+  destination: string;
+  locations: planState["locations"];
+  setDays: React.Dispatch<React.SetStateAction<planState[]>>;
+}
+export interface tripDeleteProps {
+  location: Location;
+  setDays: React.Dispatch<React.SetStateAction<planState[]>>;
+  tripId: number;
+}
+export interface LocationCardProps {
+  locationValue: Location;
+  setIsEditModalVisible: React.Dispatch<React.SetStateAction<boolean>>;
+  handleDelete: () => void;
+  renderHashtags: () => React.JSX.Element | null;
+  handleWebView: () => void;
+}
+export interface LocationValueProps {
+  locationValue: Location;
+}
